feat(api-transform): add optional key handling in data transformation
Added support for transforming data to keys that might not be present in the object through a new
flag 'shouldSetValueEvenIfMissing'. This enhances flexibility when handling API data
transformations.

Added a new error type KEY_FOR_TRANSFORM_NOT_IN_OBJECT to provide better error handling
when a key specified for transformation is not found in the target object.